Updated with exception handeling when generating message
When creating the message in SoapMessageEncoder at line 132 "Message.CreateMessage(reader, maxSizeOfHeaders, MessageVersion)", it can return in a exception which isn't catch.
This can be if the SOAP body looks like this
"Something completely different"
Hence by having the try - catch block we make sure the exception is returned and being picked up by the implementation of the IFaultExceptionTransFormer.