Fix bug ArrayOfString maxOccurs="1" bug
This is a regression of #960.
In commit 5026da3f41660ab732b6aa5491a17d0aef33c17e, the AddSchemaType()
method was changed to contain a special case for string and byte[].
However this breaks the situation where the inner type of a list is
string. In that case, maxOccurs must be set to "unbounded" for the
string element. So this commit reverts that change and instead fixes the
original issue at another place.

Fixes #988