WSDL for XmlSerializer: Fix invalid ArrayOfString schema
Removed the special if-condition for "string", which does not make sense,
because it does not correspond to the actual XML that is generated
by the XmlSerializer. Afterwards, the fields _arrayToBuild and
_buildArrayTypes are not needed anymore, so I have removed them.

The test CheckStringArrayNameWsdl() can only be performed for the
DataContractSerializer now, because the complexType does not have an
element inside it anymore.

Fixes #996