XmlSerializer WSDL: XmlAttribute: Set use="required" attribute
If a member has been annotated with the [XmlAttribute] attribute, and it
is a value type, such as int, it will get the use="required" attribute.
However if it has a ShouldSerialize*() method, then it is considered
optional and the use="required" attribute is not added.

Added a new test to verify this behavior.

Fixes #1006