Do not use chunked transfer encoding on ASP.NET Core 2.1
Setting the Content-Length header explicitly will disable chunked
transfer-encoding. However this was only done for one of the two
WriteMessageAsync methods in SoapMessageEncoder. I have changed the
other method to also set the Content-Length header.

To do this I have used a MemoryStream. The other implementation used
a StringWriter, which is not ideal, because it will convert to and from
different encodings. We only want to find out the length of the stream,
so a MemoryStream is faster and sufficient. Also changed the other
WriteMessageAsync implementation to use a MemoryStream and removed the
CustomStringWriter class.

Fixes #1015