Add option to generate soapAction without contract name
By default, the soapAction that is generated if it is not specified in
the [OperationContract(Action = "...")] attribute, follows the format
{namespace}/{contractName}/{methodName}.

This seems to be different than the default soapAction that is
generated in ASMX web services in .NET Framework. The new option makes
it possible to omit the contractName and instead generate
soapActions with the format {namespace}/{methodName}. The default for
this new option is false. This means that the behavior for existing
users of SoapCore does not change.